Modify CSNichrome for async automatic shutoff
CSNichrome now utilizes an update() function that guarantees that
the nichrome cutting circuit never runs for more than 10 seconds.
With simple modification, this can be made to be configurable.

When you want to cut, call start(). When you want to stop, call
stop(). If you don't call stop() in time, then it will do it for you.
You must therefore keep track of time on your own.