Modify BME lib to use local Adafruit Sensor header
In order to allow easy cross-compilation across multiple machines,
the BME library header file has been modified to use the local version
of the Adafruit Sensor unified header. This would mean that you don't
have to touch the `Adafruit/libraries` directory in order to compile
the code on your own machine. Further, it means that changes to the
sensor header are now fully tracked in the git history of this repo.